Implement login on database
- Implement Packet_Character
- Housekeeping

+++ managers/databasemanager.cpp
@@ -4,15 +4,20 @@
 #include <iostream>
 #include <format>
 #include "serverconfig.h"
+#include "libbcrypt/BCrypt.hpp"
 
 DatabaseManager databaseManager;
 
 DatabaseManager::DatabaseManager() {
     _connection = NULL;
+    _addedServerChannel = false;
 }
 
 DatabaseManager::~DatabaseManager() {
+    databaseManager.RemoveServerChannel();
+
     mysql_close(_connection);
+    _connection = NULL;
 }
 
 bool DatabaseManager::Init(string server, string user, string password, string database) {
@@ -57,16 +62,21 @@ bool DatabaseManager::AddServerChannel() {
         return false;
     }
 
+    _addedServerChannel = true;
     return true;
 }
 
 void DatabaseManager::RemoveServerChannel() {
-    string query;
-    query = format("DELETE FROM server_channels WHERE channelID = {} AND serverID = {};", serverConfig.channelID, serverConfig.serverID);
+    if (_addedServerChannel) {
+        string query;
+        query = format("DELETE FROM server_channels WHERE channelID = {} AND serverID = {};", serverConfig.channelID, serverConfig.serverID);
 
-    if (mysql_query(_connection, query.c_str())) {
-        cout << format("[DatabaseManager] Query error on RemoveServerChannel: {}\n", mysql_error(_connection));
-        return;
+        if (mysql_query(_connection, query.c_str())) {
+            cout << format("[DatabaseManager] Query error on RemoveServerChannel: {}\n", mysql_error(_connection));
+            return;
+        }
+
+        _addedServerChannel = false;
     }
 }
 
@@ -81,9 +91,12 @@ void DatabaseManager::UpdateChannelNumPlayers(unsigned short numPlayers) {
 }
 
 void DatabaseManager::GetChannelsNumPlayers() {
+    string query;
+    MYSQL_RES* res = NULL;
+    MYSQL_ROW row = NULL;
+
     for (auto& server : serverConfig.serverList) {
         if (server.id == serverConfig.serverID) {
-            string query;
             query = format("SELECT channelID, numPlayers FROM server_channels WHERE serverID = {};", server.id);
 
             if (mysql_query(_connection, query.c_str())) {
@@ -91,8 +104,7 @@ void DatabaseManager::GetChannelsNumPlayers() {
                 return;
             }
 
-            MYSQL_RES* res = mysql_use_result(_connection);
-            MYSQL_ROW row;
+            res = mysql_use_result(_connection);
 
             while ((row = mysql_fetch_row(res)) != NULL)
             {
@@ -108,7 +120,6 @@ void DatabaseManager::GetChannelsNumPlayers() {
             mysql_free_result(res);
         }
         else {
-            string query;
             query = format("SELECT 1 FROM server_channels WHERE serverID = {};", server.id);
 
             if (mysql_query(_connection, query.c_str())) {
@@ -116,7 +127,7 @@ void DatabaseManager::GetChannelsNumPlayers() {
                 return;
             }
 
-            MYSQL_RES* res = mysql_use_result(_connection);
+            res = mysql_use_result(_connection);
 
             if (mysql_fetch_row(res) == NULL) {
                 mysql_free_result(res);
@@ -128,7 +139,6 @@ void DatabaseManager::GetChannelsNumPlayers() {
 
                 server.status = ServerStatus::Ready;
 
-                string query;
                 query = format("SELECT channelID, numPlayers FROM server_channels WHERE serverID = {};", server.id);
 
                 if (mysql_query(_connection, query.c_str())) {
@@ -136,8 +146,7 @@ void DatabaseManager::GetChannelsNumPlayers() {
                     return;
                 }
 
-                MYSQL_RES* res = mysql_use_result(_connection);
-                MYSQL_ROW row;
+                res = mysql_use_result(_connection);
 
                 while ((row = mysql_fetch_row(res)) != NULL)
                 {
@@ -152,6 +161,58 @@ void DatabaseManager::GetChannelsNumPlayers() {
     }
 }
 
-bool DatabaseManager::Login(string userName, string password) {
-    return true;
+Packet_ReplyType DatabaseManager::Login(string userName, string password) {
+    if (userName.empty()) {
+        return Packet_ReplyType::InvalidName;
+    }
+
+    if (password.empty()) {
+        return Packet_ReplyType::InvalidPassword;
+    }
+
+    string query;
+    query = format("SELECT userID, password FROM users WHERE userName = '{}';", userName);
+
+    if (mysql_query(_connection, query.c_str())) {
+        cout << format("[DatabaseManager] Query error on Login: {}\n", mysql_error(_connection));
+        return Packet_ReplyType::SysError;
+    }
+
+    MYSQL_RES* res = mysql_use_result(_connection);
+    MYSQL_ROW row = mysql_fetch_row(res);
+
+    Packet_ReplyType result = Packet_ReplyType::LoginSuccess;
+
+    if (row != NULL) {
+        cout << format("password: {}, hash: {}\n", password.c_str(), row[1]);
+
+        if (!BCrypt::validatePassword(password, row[1])) {
+            mysql_free_result(res);
+
+            return Packet_ReplyType::WrongPassword;
+        }
+
+        query = format("SELECT 1 FROM game_users WHERE userID = {};", row[0]);
+
+        mysql_free_result(res);
+
+        if (mysql_query(_connection, query.c_str())) {
+            cout << format("[DatabaseManager] Query error on Login: {}\n", mysql_error(_connection));
+            return Packet_ReplyType::SysError;
+        }
+
+        res = mysql_use_result(_connection);
+        row = mysql_fetch_row(res);
+
+        if (row == NULL) {
+            result = Packet_ReplyType::CreateCharacter;
+        }
+    }
+    else {
+        result = Packet_ReplyType::NotExist;
+    }
+
+    mysql_free_result(res);
+
+    return result;
 }
